requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Real-time Streaming
pathway==0.7.7

# Redis
redis==5.0.1
hiredis==2.2.3

# Machine Learning
xgboost==2.0.2
lightgbm==4.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# LLM Integration
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and Monitoring
structlog==23.2.0